requirements.txt
fastapi
uvicorn
pillow
pypotrace
python-multipart
